Co-locate SeverityFormatter with SeverityLevel and clean up stale post-refactor comments
